USAGOV-2779-tome-intrumentation-and-ssg-overhead: refacorization and optimization of TomeStaticEvents and DOM parsing calls
