--- web/modules/custom/usagov_ssg_postprocessing/src/EventSubscriber/TomeEventSubscriber.php
@@ -1,477 +1,0 @@
-<?php
-
-namespace Drupal\usagov_ssg_postprocessing\EventSubscriber;
-
-use Drupal\Core\Entity\ContentEntityBase;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
-use Drupal\Core\Language\LanguageManagerInterface;
-use Drupal\Core\Site\Settings;
-use Drupal\tome_static\Event\CollectPathsEvent;
-use Drupal\tome_static\Event\ModifyHtmlEvent;
-use Drupal\tome_static\Event\PathPlaceholderEvent;
-use Drupal\tome_static\Event\TomeStaticEvents;
-use Drupal\usagov_ssg_postprocessing\SsgMetricTrait;
-use Drupal\views\ViewExecutable;
-use Drupal\views\Views;
-use Masterminds\HTML5;
-use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-
-/**
- * This event subscriber modifies static site generation:
- *  - During path collection, removes excluded directories -- allowing us to
- *    specify the omission of entire directories like jsonapi, node, etc.
- *  - During HTML modification, adds a trailing slash to links to the "/es" path.
- *
- * @internal
- */
-class TomeEventSubscriber implements EventSubscriberInterface {
-
-  use SsgMetricTrait;
-
-  /**
-   * The entity type manager.
-   *
-   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
-   */
-  protected $entityTypeManager;
-
-  /**
-   * The language manager.
-   *
-   * @var \Drupal\Core\Language\LanguageManagerInterface
-   */
-  protected $languageManager;
-
-  /**
-   * Constructs the EntityPathSubscriber object.
-   *
-   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
-   *   The entity type manager.
-   * @param \Drupal\Core\Language\LanguageManagerInterface $language_manager
-   *   The language manager.
-   */
-  public function __construct(EntityTypeManagerInterface $entity_type_manager, LanguageManagerInterface $language_manager) {
-    $this->entityTypeManager = $entity_type_manager;
-    $this->languageManager = $language_manager;
-  }
-
-  /**
-   * Reacts to a collect paths event. Excludes entire directories by
-   * deleting any paths that match the specified string,
-   * or that start with the string and a /.
-   * Such directories come from the setting usagov_tome_static_path_exclude_directories.
-   *
-   * (To exclude individual paths, set tome_static_path_exclude -- it's built in.)
-   *
-   * @param \Drupal\tome_static\Event\CollectPathsEvent $event
-   *   The collect paths event.
-   */
-  public function excludeDirectories(CollectPathsEvent $event): void {
-    $metric_start = $this->ssgMetricStart();
-    $excluded_directories = self::getExcludedDirectories();
-    $paths = $event->getPaths(TRUE);
-    $initial_count = count($paths);
-    $placeholder_count = 0;
-    $excluded_count = 0;
-    foreach ($paths as $path => $metadata) {
-      /**
-       * We are going to spend the time here to get the "real" paths for any
-       * placeholder-ed paths, so we can identify and exclude what we want to skip.
-       * Tome would normally do this later in its process.
-       */
-      $path_parts = explode(':', $path);
-      if ($path_parts[0] == '_entity') {
-        $placeholder_count++;
-        $entity_type = $path_parts[1];
-        $langcode = $path_parts[2];
-        $entity_id = $path_parts[3];
-
-        $entity = $this->entityTypeManager->getStorage($entity_type)->load($entity_id);
-        // ContentEntityBase interface require the getTranslation()/hasTranslation() methods
-        if (!$entity || (!$entity instanceof ContentEntityBase) || !$entity->hasTranslation($langcode)) {
-          continue;
-        }
-        $entity = $entity->getTranslation($langcode);
-        $url = $entity->toUrl('canonical');
-        if (!$entity->access('view') || ($entity->isDefaultTranslation() && !$url->access())) {
-          continue;
-        }
-        if ($newpath = parse_url($url->toString(), PHP_URL_PATH)) {
-          unset($paths[$path]);
-          $metadata['original_path'] = $path;
-          $path = $newpath; // Next block tests $path against excluded directories.
-          $paths[$path] = $metadata;
-        }
-      }
-      // Get the base path without trailing slash if we're exporting to a sub-directory
-      $base_path = rtrim(trim(base_path()), '/');
-      foreach ($excluded_directories as $excluded_directory_path) {
-        $excluded_directory = $excluded_directory_path . '/';
-        if (($path == $excluded_directory_path) ||
-            ($path == $base_path . $excluded_directory_path) ||
-            (isset($metadata['original_path']) && ($metadata['original_path'] == $excluded_directory_path))) {
-          if (isset($paths[$path])) {
-            unset($paths[$path]);
-            $excluded_count++;
-          }
-        }
-        elseif (str_starts_with($path, $excluded_directory) ||
-            (str_starts_with($path, $base_path . $excluded_directory)) ||
-            (isset($metadata['original_path']) && str_starts_with($metadata['original_path'], $excluded_directory))) {
-          if (isset($paths[$path])) {
-            unset($paths[$path]);
-            $excluded_count++;
-          }
-        }
-      }
-    }
-    $event->replacePaths($paths);
-    $this->ssgMetricEnd('tome_exclude_directories', $metric_start, 'end', [
-      'initial_count' => $initial_count,
-      'final_count' => count($paths),
-      'excluded_count' => $excluded_count,
-      'placeholder_count' => $placeholder_count,
-      'excluded_directory_count' => count($excluded_directories),
-    ]);
-  }
-
-  /**
-   * Returns per-site excluded directory paths.
-   *
-   * @return array<mixed>
-   *   An array of excluded paths.
-   */
-  public static function getExcludedDirectories(): array {
-    $excluded_paths = [];
-    $site_paths = Settings::get('usagov_tome_static_path_exclude_directories', []);
-    if (is_array($site_paths)) {
-      foreach ($site_paths as $path) {
-        $excluded_paths[] = $path;
-      }
-    }
-    return $excluded_paths;
-  }
-
-  /**
-   * Reacts to a modify HTML event; replacing links to /es (possibly with a query or
-   * fragment appended) with /es/.
-   *
-   * @param \Drupal\tome_static\Event\ModifyHtmlEvent $event
-   *   The event.
-   */
-  public function modifyHtml(ModifyHtmlEvent $event): void {
-    $metric_start = $this->ssgMetricStart();
-    $html = $event->getHtml();
-    $html5 = new HTML5();
-
-    // QuickFix for USAGOV-2312. We want to remove this meta-data.
-    $html = str_replace('xmlns:xlink="http://www.w3.org/1999/xlink"', '', $html);
-    $event->setHtml($html); // just in case setHtml() dosnt get caled later
-
-    // LIBXML_SCHEMA_CREATE fixes a problem wherein DOMDocument would remove closing HTML
-    // tags within quoted text in a script element. See https://bugs.php.net/bug.php?id=74628
-    $document = new \DOMDocument();
-    @$document->loadHTML($html, LIBXML_SCHEMA_CREATE);
-
-    $xpath = new \DOMXPath($document);
-    $changes = FALSE;
-
-    // --- USAGOV-2515: Replace spaces with '+' in image/media src and srcset attributes ---
-    // img[src], source[src], and any srcset attributes
-    $img_nodes = $xpath->query('//*[@src]');
-    $image_node_count = $img_nodes ? $img_nodes->length : 0;
-    $image_src_rewrite_count = 0;
-    $debug_image_paths = $this->debugImagePathsEnabled();
-    foreach ($img_nodes as $node) {
-      if ($node instanceof \DOMElement) {
-        $src = $node->getAttribute('src');
-        if (strpos($src, ' ') !== FALSE || strpos($src, '%20') !== FALSE) {
-          $new_src = str_replace([' ', '%20'], '+', $src);
-          $node->setAttribute('src', $new_src);
-          $changes = TRUE;
-          $image_src_rewrite_count++;
-          if ($debug_image_paths) {
-            file_put_contents('/tmp/tome-img-path-debug.log', '[IMG] Replaced src: ' . $src . ' -> ' . $new_src . "\n", FILE_APPEND);
-          }
-        }
-        elseif ($debug_image_paths) {
-          file_put_contents('/tmp/tome-img-path-debug.log', '[IMG] No change src: ' . $src . "\n", FILE_APPEND);
-        }
-      }
-    }
-
-    // Existing logic: fix /es links
-    $nodes = $xpath->query('//a[starts-with(@href,"/es")]');
-    $es_link_count = $nodes ? $nodes->length : 0;
-    $es_link_rewrite_count = 0;
-
-    /** @var \DOMElement $node */
-    foreach ($nodes as $node) {
-      $original_href = $node->getAttribute('href');
-      $new_href = NULL;
-
-      if ($original_href === '/es') {
-        $new_href = '/es/';
-      }
-      elseif ($original_href[3] === '?') {
-        $new_href = '/es/' . substr($original_href, 3);
-      }
-      elseif ($original_href[3] === '#') {
-        $new_href = '/es/' . substr($original_href, 3);
-      }
-      if ($new_href) {
-        $changes = TRUE;
-        $es_link_rewrite_count++;
-        $node->setAttribute('href', $new_href);
-        $event->addExcludePath($new_href);
-      }
-    }
-
-    // Never crawl the rewritten Spanish path. It might be treated like a redirect by
-    // Tome and overwrite the original homepage HTML
-    $event->addExcludePath('/es/');
-
-    if ($changes) {
-      // Render it as HTML5:
-      $modifiedHtml = $html5->saveHTML($document);
-      // QuickFix for USAGOV-2312. We want to remove this meta-data.
-      $modifiedHtml = str_replace(
-        'xmlns:xlink="http://www.w3.org/1999/xlink"',
-        '',
-        $modifiedHtml
-      );
-      $event->setHtml($modifiedHtml);
-    }
-    $this->ssgMetricEnd('tome_modify_html', $metric_start, 'end', [
-      'path' => $event->getPath(),
-      'changed' => $changes,
-      'html_bytes' => strlen($html),
-      'image_node_count' => $image_node_count,
-      'image_src_rewrite_count' => $image_src_rewrite_count,
-      'debug_image_paths' => $debug_image_paths,
-      'es_link_count' => $es_link_count,
-      'es_link_rewrite_count' => $es_link_rewrite_count,
-    ]);
-  }
-
-  protected function debugImagePathsEnabled(): bool {
-    return filter_var(getenv('SSG_DEBUG_IMAGE_PATHS') ?: FALSE, FILTER_VALIDATE_BOOLEAN);
-  }
-
-  /**
-   * Prevent exporting paths Tome might discover after the collect paths event.
-   *
-   * @param PathPlaceholderEvent $event
-   * @return void
-   */
-  public function excludeInvalidPaths(PathPlaceholderEvent $event) {
-    $metric_start = $this->ssgMetricStart();
-    $path = $event->getPath();
-
-    if ($path !== '/' && str_ends_with($path, '/')) {
-      // Tome should never request the Spanish homepage or any other local path
-      // with a trailing-slash. If it does request it, that is because the path
-      // was found in the content of a node or term.
-      // For example, when tome runs and it finds a link to `/es/`, Drupal will
-      // redirect the request for `/es/` to `/es`. The response causes Tome to
-      // save it in  the contents of `es/index.html` with an refresh redirect.
-      $event->setInvalid();
-      $this->ssgMetricEnd('tome_path_placeholder_filter', $metric_start, 'invalid', [
-        'path' => $path,
-        'reason' => 'trailing_slash',
-      ]);
-      return;
-    }
-
-    if (preg_match('/(es\/)?node\/\d+$/', $path)) {
-      $event->setInvalid();
-      $this->ssgMetricEnd('tome_path_placeholder_filter', $metric_start, 'invalid', [
-        'path' => $path,
-        'reason' => 'node_path',
-      ]);
-      return;
-    }
-
-    $this->ssgMetricEnd('tome_path_placeholder_filter', $metric_start, 'valid', [
-      'path' => $path,
-    ]);
-  }
-
-  /**
-   * Add agency index paths to be exported instead of relying on Tome discovering the path
-   */
-  public function addAgencyIndexes(CollectPathsEvent $event): void {
-    $metric_start = $this->ssgMetricStart();
-    $metadata = ['language_processed' => TRUE];
-    // Get the English letters to output from the pager view
-    $view = Views::getView('federal_agencies');
-    $view->setDisplay('attachment_1');
-
-    $metadata['langcode'] = 'en';
-    $english_letters = $this->getLetters($view);
-    foreach ($english_letters as $letter) {
-      $event->addPath('/agency-index?letter=' . $letter, $metadata);
-    }
-
-    // Get the Spanish letters
-    $view = Views::getView('federal_agencies');
-
-    $view->setDisplay('attachment_2');
-
-    $metadata['langcode'] = 'es';
-    $spanish_letters = $this->getLetters($view);
-    foreach ($spanish_letters as $letter) {
-      $event->addPath('/es/indice-agencias?letter=' . $letter, $metadata);
-    }
-
-    $this->ssgMetricEnd('tome_add_agency_indexes', $metric_start, 'end', [
-      'english_count' => count($english_letters),
-      'spanish_count' => count($spanish_letters),
-    ]);
-
-  }
-
-  /**
-   * Clear the cache for the blog listing. This isn't "collecting paths," but
-   * it is a thing that we want to get done once before we start processing
-   * the blog pages.
-   *
-   * @param CollectPathsEvent $event
-   * @return void
-   */
-  public function clearBlogViewCache(CollectPathsEvent $event): void {
-    $metric_start = $this->ssgMetricStart();
-    $view = Views::getView('blog_menu');
-    $view->storage->invalidateCaches();
-    $this->ssgMetricEnd('tome_clear_blog_view_cache', $metric_start);
-  }
-
-  /**
-   * Add blog year and month archive paths to be exported.
-   *
-   * The blog view has optional contextual filters (year/month) which means
-   * Tome doesn't automatically discover the base /blog path. We explicitly
-   * add it here to ensure /blog is generated.
-
-   *
-   * Also, year and month archive pages (e.g. /blog/2018, /blog/2018/03) are served
-   * by the blog view with contextual filters, so Tome cannot reliably discover them
-   * automatically. We read the paths directly from the usagov_blog_menu
-   * menu link entities — which are only created for years/months that have
-   * published content — and register them explicitly.
-   */
-  public function addBlogPaths(CollectPathsEvent $event): void {
-    $metric_start = $this->ssgMetricStart();
-    $metadata = ['language_processed' => TRUE, 'langcode' => 'en'];
-    $added_count = 0;
-
-    // Add the base /blog path.
-    $event->addPath('/blog', $metadata);
-    $added_count++;
-
-    $links = $this->entityTypeManager
-      ->getStorage('menu_link_content')
-      ->loadByProperties(['menu_name' => 'usagov_blog_menu', 'enabled' => 1]);
-
-    foreach ($links as $link) {
-      /** @var \Drupal\menu_link_content\Entity\MenuLinkContent $link */
-      $uri = $link->get('link')->uri ?? '';
-      // URIs are in the form "internal:/blog/YYYY" or "internal:/blog/YYYY/MM".
-      // Skip node links (individual blog posts) which Tome discovers normally.
-      if (!str_starts_with($uri, 'internal:/blog/')) {
-        continue;
-      }
-      $path = substr($uri, strlen('internal:'));
-      // Only register year (/blog/YYYY) and month (/blog/YYYY/MM) paths.
-      if (preg_match('#^/blog/\d{4}(/\d{2})?$#', $path)) {
-        $event->addPath($path, $metadata);
-        $added_count++;
-      }
-    }
-    $this->ssgMetricEnd('tome_add_blog_paths', $metric_start, 'end', [
-      'added_count' => $added_count,
-      'menu_link_count' => count($links),
-    ]);
-  }
-
-  /**
-   * Logs aggregate path counts after custom and contrib path collection.
-   */
-  public function logCollectedPathCounts(CollectPathsEvent $event): void {
-    $metric_start = $this->ssgMetricStart();
-    $paths = $event->getPaths(TRUE);
-    $counts = [
-      'total_count' => count($paths),
-      'entity_placeholder_count' => 0,
-      'benefit_search_count' => 0,
-      'agency_index_count' => 0,
-      'blog_count' => 0,
-      'json_count' => 0,
-      'english_count' => 0,
-      'spanish_count' => 0,
-      'language_processed_count' => 0,
-    ];
-
-    foreach ($paths as $path => $metadata) {
-      if (str_starts_with($path, '_entity:')) {
-        $counts['entity_placeholder_count']++;
-      }
-      if (str_contains($path, '/_data/benefits-search/')) {
-        $counts['benefit_search_count']++;
-      }
-      if (str_starts_with($path, '/agency-index') || str_starts_with($path, '/es/indice-agencias')) {
-        $counts['agency_index_count']++;
-      }
-      if ($path === '/blog' || str_starts_with($path, '/blog/')) {
-        $counts['blog_count']++;
-      }
-      if (str_ends_with(parse_url($path, PHP_URL_PATH) ?? $path, '.json')) {
-        $counts['json_count']++;
-      }
-      if (($metadata['langcode'] ?? NULL) === 'en') {
-        $counts['english_count']++;
-      }
-      if (($metadata['langcode'] ?? NULL) === 'es') {
-        $counts['spanish_count']++;
-      }
-      if (!empty($metadata['language_processed'])) {
-        $counts['language_processed_count']++;
-      }
-    }
-
-    $this->ssgMetricEnd('tome_collected_path_counts', $metric_start, 'end', $counts);
-  }
-
-  /**
-   * @return string[]
-   */
-  private function getLetters(ViewExecutable $view): array {
-    $view->execute();
-    $letters = [];
-    foreach ($view->result as $result) {
-      // Tome must create folders that are lower-cased
-      $letter = strtolower($result->title_truncated);
-      // The "A" page is the default agency-index page, no need to export.
-      if ($letter !== 'a') {
-        $letters[] = $letter;
-      }
-    }
-
-    return array_unique($letters);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function getSubscribedEvents() {
-    $events[TomeStaticEvents::MODIFY_HTML][] = ['modifyHtml'];
-    $events[TomeStaticEvents::COLLECT_PATHS][] = ['excludeDirectories'];
-    $events[TomeStaticEvents::COLLECT_PATHS][] = ['addAgencyIndexes'];
-    $events[TomeStaticEvents::COLLECT_PATHS][] = ['clearBlogViewCache'];
-    $events[TomeStaticEvents::COLLECT_PATHS][] = ['addBlogPaths'];
-    $events[TomeStaticEvents::COLLECT_PATHS][] = ['logCollectedPathCounts', -1000];
-    $events[TomeStaticEvents::PATH_PLACEHOLDER][] = ['excludeInvalidPaths'];
-    return $events;
-  }
-
-}
