+++ web/modules/custom/usagov_ssg_postprocessing/usagov_ssg_postprocessing.services.yml
@@ -1,11 +1,19 @@
 services:
-  usagov_ssg_postprocessing_tome_misc_events_subscriber:
-    class: '\Drupal\usagov_ssg_postprocessing\EventSubscriber\TomeEventSubscriber'
-    arguments: ['@entity_type.manager', '@language_manager']
+  usagov_ssg_postprocessing.static_html_subscriber:
+    class: '\Drupal\usagov_ssg_postprocessing\EventSubscriber\StaticHtmlSubscriber'
     tags:
       - { name: 'event_subscriber' }
-  usagov_ssg_postprocessing_pagerpath_events_subscriber:
-    class: '\Drupal\usagov_ssg_postprocessing\EventSubscriber\PagerPathSubscriber'
+  usagov_ssg_postprocessing.static_path_collection_subscriber:
+    class: '\Drupal\usagov_ssg_postprocessing\EventSubscriber\StaticPathCollectionSubscriber'
+    arguments: ['@entity_type.manager']
+    tags:
+      - { name: 'event_subscriber' }
+  usagov_ssg_postprocessing.static_path_placeholder_subscriber:
+    class: '\Drupal\usagov_ssg_postprocessing\EventSubscriber\StaticPathPlaceholderSubscriber'
+    tags:
+      - { name: 'event_subscriber' }
+  usagov_ssg_postprocessing.static_destination_subscriber:
+    class: '\Drupal\usagov_ssg_postprocessing\EventSubscriber\StaticDestinationSubscriber'
     tags:
       - { name: 'event_subscriber' }
   usagov_ssg_postprocessing_request_prepare_events_subscriber:
